Improve `get_steam_games`
get_steam_games was previously finding the `UnityCrashHandler` exes meaning they were not closed.

This has 2 improvements:

Firstly remove `steam_path` from the config as that is now found from the registry
Secondly, parse the appmanifest files to get the game install paths. This allows for the game to be found if it is installed on a different drive to the Steam install.